_archive_v2/01_Fundamentals/requirements.txt
google-genai
httpx
python-dotenv
